+++ lib/models/additionalWeatherData.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 class AdditionalWeatherData {
   final String precipitation;
-  final int uvi;
+  final double uvi;
   final int clouds;
   AdditionalWeatherData({
     required this.precipitation,
@@ -15,7 +15,7 @@ class AdditionalWeatherData {
     final precipitation = calcPrecip.toStringAsFixed(0);
     return AdditionalWeatherData(
       precipitation: precipitation,
-      uvi: (json['daily'][0]['uvi']).toInt(),
+      uvi: (json['daily'][0]['uvi']).toDouble(),
       clouds: json['daily'][0]['clouds'] ?? 0,
     );
   }
